Adding JPA Delete Customer by Id Not found

+++ spring-6-rest-mvc/src/main/java/com/zlateva/spring6restmvc/services/CustomerServiceJPA.java
@@ -54,8 +54,12 @@ public class CustomerServiceJPA implements CustomerService {
     }
 
     @Override
-    public void deleteCustomerById(UUID customerId) {
-        customerRepository.deleteById(customerId);
+    public Boolean deleteCustomerById(UUID customerId) {
+        if (customerRepository.existsById(customerId)) {
+            customerRepository.deleteById(customerId);
+            return true;
+        }
+        return false;
     }
 
     @Override
